refactor(rpc-server): Tweaks and clean up around query.call_function method implementation (#258)
* refactor(rpc-server): Tweaks and clean up around query.call_function method implementation

* Follow the clippy suggestion

* Drop redundant pieces of code
