+++ rpc-server/src/modules/blocks/utils.rs
@@ -78,16 +78,16 @@ pub async fn fetch_chunk_from_s3(
 #[cfg_attr(feature = "tracing-instrumentation", tracing::instrument(skip(data)))]
 pub async fn fetch_block_from_cache_or_get(
     data: &jsonrpc_v2::Data<ServerContext>,
-    block_reference: near_primitives::types::BlockReference,
+    block_reference: &near_primitives::types::BlockReference,
     method_name: &str,
 ) -> Result<CacheBlock, near_jsonrpc::primitives::types::blocks::RpcBlockError> {
-    let block = match block_reference.clone() {
+    let block = match block_reference {
         near_primitives::types::BlockReference::BlockId(block_id) => {
             let block_height = match block_id {
-                near_primitives::types::BlockId::Height(block_height) => block_height,
+                near_primitives::types::BlockId::Height(block_height) => *block_height,
                 near_primitives::types::BlockId::Hash(hash) => data
                     .db_manager
-                    .get_block_by_hash(hash, method_name)
+                    .get_block_by_hash(*hash, method_name)
                     .await
                     .map_err(|err| {
                         near_jsonrpc::primitives::types::blocks::RpcBlockError::UnknownBlock {
