+++ rpc-server/src/modules/state/methods.rs
@@ -9,7 +9,7 @@ pub async fn view_state_paginated(
     Params(params): Params<crate::modules::state::RpcViewStatePaginatedRequest>,
 ) -> Result<crate::modules::state::RpcViewStatePaginatedResponse, RPCError> {
     let block_reference = near_primitives::types::BlockReference::BlockId(params.block_id.clone());
-    let block = fetch_block_from_cache_or_get(&data, block_reference, "view_state_paginated")
+    let block = fetch_block_from_cache_or_get(&data, &block_reference, "view_state_paginated")
         .await
         .map_err(near_jsonrpc::primitives::errors::RpcError::from)?;
